requirements.txt
openai
streamlit
chromadb
langchain
langchain-community
langchain-openai 
langchain-chroma
tiktoken
st-pages
python-dotenv
PyPDF2
faiss-cpu


-e .